+++ back/src/test/kotlin/com/back/architecture/ArchitectureGuardTest.kt
@@ -175,7 +175,13 @@ class ArchitectureGuardTest {
     fun `핵심 web controller는 application service 구현체를 직접 참조하지 않아야 한다`() {
         noClasses()
             .that()
-            .haveFullyQualifiedName("com.back.boundedContexts.post.adapter.web.ApiV1PostController")
+            .haveFullyQualifiedName("com.back.boundedContexts.post.adapter.web.ApiV1PostPublicReadController")
+            .or()
+            .haveFullyQualifiedName("com.back.boundedContexts.post.adapter.web.ApiV1PostCommandController")
+            .or()
+            .haveFullyQualifiedName("com.back.boundedContexts.post.adapter.web.ApiV1PostInteractionController")
+            .or()
+            .haveFullyQualifiedName("com.back.boundedContexts.post.adapter.web.ApiV1PostDraftController")
             .or()
             .haveFullyQualifiedName("com.back.boundedContexts.post.adapter.web.ApiV1AdmPostController")
             .or()
